@@ -1,20 +1,17 @@
-﻿using System.Collections.Generic;
-using System.Diagnostics;
-using System.Linq;
+﻿using System.Diagnostics;
 
-namespace ConsoleDI.IEnumerableExample
+namespace ConsoleDI.IEnumerableExample;
+
+public class ExampleService
 {
-    public class ExampleService
+    public ExampleService(
+        IMessageWriter messageWriter,
+        IEnumerable<IMessageWriter> messageWriters)
     {
-        public ExampleService(
-            IMessageWriter messageWriter,
-            IEnumerable<IMessageWriter> messageWriters)
-        {
-            Trace.Assert(messageWriter is LoggingMessageWriter);
+        Trace.Assert(messageWriter is LoggingMessageWriter);
 
-            var dependencyArray = messageWriters.ToArray();
-            Trace.Assert(dependencyArray[0] is ConsoleMessageWriter);
-            Trace.Assert(dependencyArray[1] is LoggingMessageWriter);
-        }
+        var dependencyArray = messageWriters.ToArray();
+        Trace.Assert(dependencyArray[0] is ConsoleMessageWriter);
+        Trace.Assert(dependencyArray[1] is LoggingMessageWriter);
     }
 }
