@@ -1,15 +1,14 @@
 ﻿using Microsoft.Extensions.Logging;
 
-namespace ConsoleDI.IEnumerableExample
+namespace ConsoleDI.IEnumerableExample;
+
+public class LoggingMessageWriter : IMessageWriter
 {
-    public class LoggingMessageWriter : IMessageWriter
-    {
-        private readonly ILogger<LoggingMessageWriter> _logger;
+    private readonly ILogger<LoggingMessageWriter> _logger;
 
-        public LoggingMessageWriter(ILogger<LoggingMessageWriter> logger) =>
-            _logger = logger;
+    public LoggingMessageWriter(ILogger<LoggingMessageWriter> logger) =>
+        _logger = logger;
 
-        public void Write(string message) =>
-            _logger.LogInformation(message);
-    }
+    public void Write(string message) =>
+        _logger.LogInformation(message);
 }
