@@ -1,12 +1,11 @@
 ﻿using static System.Guid;
 
-namespace ConsoleDI.Example
+namespace ConsoleDI.Example;
+
+public class DefaultOperation :
+    ITransientOperation,
+    IScopedOperation,
+    ISingletonOperation
 {
-    public class DefaultOperation :
-        ITransientOperation,
-        IScopedOperation,
-        ISingletonOperation
-    {
-        public string OperationId { get; } = NewGuid().ToString()[^4..];
-    }
+    public string OperationId { get; } = NewGuid().ToString()[^4..];
 }
