@@ -1,60 +1,57 @@
-﻿using System;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
+﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace ConsoleDI.Example
+namespace ConsoleDI.Example;
+
+class Program
 {
-    class Program
+    static Task Main(string[] args)
     {
-        static Task Main(string[] args)
-        {
-            using IHost host = CreateHostBuilder(args).Build();
+        using IHost host = CreateHostBuilder(args).Build();
 
-            ExemplifyScoping(host.Services, "Scope 1");
-            ExemplifyScoping(host.Services, "Scope 2");
+        ExemplifyScoping(host.Services, "Scope 1");
+        ExemplifyScoping(host.Services, "Scope 2");
 
-            return host.RunAsync();
-        }
-        // Sample output:
-        // Scope 1-Call 1 .GetRequiredService<OperationLogger>(): ITransientOperation [ 80f4...Always different        ]
-        // Scope 1-Call 1 .GetRequiredService<OperationLogger>(): IScopedOperation    [ c878...Changes only with scope ]
-        // Scope 1-Call 1 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
-        // ...
-        // Scope 1-Call 2 .GetRequiredService<OperationLogger>(): ITransientOperation [ f3c0...Always different        ]
-        // Scope 1-Call 2 .GetRequiredService<OperationLogger>(): IScopedOperation    [ c878...Changes only with scope ]
-        // Scope 1-Call 2 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
-        //
-        // Scope 2-Call 1 .GetRequiredService<OperationLogger>(): ITransientOperation [ f9af...Always different        ]
-        // Scope 2-Call 1 .GetRequiredService<OperationLogger>(): IScopedOperation    [ 2bd0...Changes only with scope ]
-        // Scope 2-Call 1 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
-        // ...
-        // Scope 2-Call 2 .GetRequiredService<OperationLogger>(): ITransientOperation [ fa65...Always different        ]
-        // Scope 2-Call 2 .GetRequiredService<OperationLogger>(): IScopedOperation    [ 2bd0...Changes only with scope ]
-        // Scope 2-Call 2 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
+        return host.RunAsync();
+    }
+    // Sample output:
+    // Scope 1-Call 1 .GetRequiredService<OperationLogger>(): ITransientOperation [ 80f4...Always different        ]
+    // Scope 1-Call 1 .GetRequiredService<OperationLogger>(): IScopedOperation    [ c878...Changes only with scope ]
+    // Scope 1-Call 1 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
+    // ...
+    // Scope 1-Call 2 .GetRequiredService<OperationLogger>(): ITransientOperation [ f3c0...Always different        ]
+    // Scope 1-Call 2 .GetRequiredService<OperationLogger>(): IScopedOperation    [ c878...Changes only with scope ]
+    // Scope 1-Call 2 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
+    //
+    // Scope 2-Call 1 .GetRequiredService<OperationLogger>(): ITransientOperation [ f9af...Always different        ]
+    // Scope 2-Call 1 .GetRequiredService<OperationLogger>(): IScopedOperation    [ 2bd0...Changes only with scope ]
+    // Scope 2-Call 1 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
+    // ...
+    // Scope 2-Call 2 .GetRequiredService<OperationLogger>(): ITransientOperation [ fa65...Always different        ]
+    // Scope 2-Call 2 .GetRequiredService<OperationLogger>(): IScopedOperation    [ 2bd0...Changes only with scope ]
+    // Scope 2-Call 2 .GetRequiredService<OperationLogger>(): ISingletonOperation [ 1586...Always the same         ]
 
-        static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureServices((_, services) =>
-                    services.AddTransient<ITransientOperation, DefaultOperation>()
-                            .AddScoped<IScopedOperation, DefaultOperation>()
-                            .AddSingleton<ISingletonOperation, DefaultOperation>()
-                            .AddTransient<OperationLogger>());
+    static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .ConfigureServices((_, services) =>
+                services.AddTransient<ITransientOperation, DefaultOperation>()
+                        .AddScoped<IScopedOperation, DefaultOperation>()
+                        .AddSingleton<ISingletonOperation, DefaultOperation>()
+                        .AddTransient<OperationLogger>());
 
-        static void ExemplifyScoping(IServiceProvider services, string scope)
-        {
-            using IServiceScope serviceScope = services.CreateScope();
-            IServiceProvider provider = serviceScope.ServiceProvider;
+    static void ExemplifyScoping(IServiceProvider services, string scope)
+    {
+        using IServiceScope serviceScope = services.CreateScope();
+        IServiceProvider provider = serviceScope.ServiceProvider;
 
-            OperationLogger logger = provider.GetRequiredService<OperationLogger>();
-            logger.LogOperations($"{scope}-Call 1 .GetRequiredService<OperationLogger>()");
+        OperationLogger logger = provider.GetRequiredService<OperationLogger>();
+        logger.LogOperations($"{scope}-Call 1 .GetRequiredService<OperationLogger>()");
 
-            Console.WriteLine("...");
+        Console.WriteLine("...");
 
-            logger = provider.GetRequiredService<OperationLogger>();
-            logger.LogOperations($"{scope}-Call 2 .GetRequiredService<OperationLogger>()");
+        logger = provider.GetRequiredService<OperationLogger>();
+        logger.LogOperations($"{scope}-Call 2 .GetRequiredService<OperationLogger>()");
 
-            Console.WriteLine();
-        }
+        Console.WriteLine();
     }
 }
