@@ -1,21 +1,19 @@
-﻿using System;
-using Microsoft.Extensions.Options;
+﻿using Microsoft.Extensions.Options;
 
-namespace ConsoleJson.Example
+namespace ConsoleJson.Example;
+
+public class MonitorService
 {
-    public class MonitorService
+    private readonly IOptionsMonitor<TransientFaultHandlingOptions> _monitor;
+
+    public MonitorService(IOptionsMonitor<TransientFaultHandlingOptions> monitor) =>
+        _monitor = monitor;
+
+    public void DisplayValues()
     {
-        private readonly IOptionsMonitor<TransientFaultHandlingOptions> _monitor;
+        TransientFaultHandlingOptions options = _monitor.CurrentValue;
 
-        public MonitorService(IOptionsMonitor<TransientFaultHandlingOptions> monitor) =>
-            _monitor = monitor;
-
-        public void DisplayValues()
-        {
-            TransientFaultHandlingOptions options = _monitor.CurrentValue;
-
-            Console.WriteLine($"TransientFaultHandlingOptions.Enabled={options.Enabled}");
-            Console.WriteLine($"TransientFaultHandlingOptions.AutoRetryDelay={options.AutoRetryDelay}");
-        }
+        Console.WriteLine($"TransientFaultHandlingOptions.Enabled={options.Enabled}");
+        Console.WriteLine($"TransientFaultHandlingOptions.AutoRetryDelay={options.AutoRetryDelay}");
     }
 }
