@@ -1,44 +1,41 @@
-﻿using System;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 
-namespace ConsoleJson.Example
+namespace ConsoleJson.Example;
+
+class Program
 {
-    class Program
+    static async Task Main(string[] args)
     {
-        static async Task Main(string[] args)
-        {
-            using IHost host = CreateHostBuilder(args).Build();
+        using IHost host = CreateHostBuilder(args).Build();
 
-            // Application code should start here.
+        // Application code should start here.
 
-            await host.RunAsync();
-        }
-
-        static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration((hostingContext, configuration) =>
-                {
-                    configuration.Sources.Clear();
-
-                    IHostEnvironment env = hostingContext.HostingEnvironment;
-
-                    configuration
-                        .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
-                        .AddJsonFile($"appsettings.{env.EnvironmentName}.json", true, true);
-
-                    IConfigurationRoot configurationRoot = configuration.Build();
-
-                    TransientFaultHandlingOptions options = new();
-                    configurationRoot.GetSection(nameof(TransientFaultHandlingOptions))
-                                     .Bind(options);
-
-                    Console.WriteLine($"TransientFaultHandlingOptions.Enabled={options.Enabled}");
-                    Console.WriteLine($"TransientFaultHandlingOptions.AutoRetryDelay={options.AutoRetryDelay}");
-                });
-        // Sample output:
-        //    TransientFaultHandlingOptions.Enabled=True
-        //    TransientFaultHandlingOptions.AutoRetryDelay=00:00:07
+        await host.RunAsync();
     }
+
+    static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .ConfigureAppConfiguration((hostingContext, configuration) =>
+            {
+                configuration.Sources.Clear();
+
+                IHostEnvironment env = hostingContext.HostingEnvironment;
+
+                configuration
+                    .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
+                    .AddJsonFile($"appsettings.{env.EnvironmentName}.json", true, true);
+
+                IConfigurationRoot configurationRoot = configuration.Build();
+
+                TransientFaultHandlingOptions options = new();
+                configurationRoot.GetSection(nameof(TransientFaultHandlingOptions))
+                                 .Bind(options);
+
+                Console.WriteLine($"TransientFaultHandlingOptions.Enabled={options.Enabled}");
+                Console.WriteLine($"TransientFaultHandlingOptions.AutoRetryDelay={options.AutoRetryDelay}");
+            });
+    // Sample output:
+    //    TransientFaultHandlingOptions.Enabled=True
+    //    TransientFaultHandlingOptions.AutoRetryDelay=00:00:07
 }
