@@ -1,19 +1,17 @@
-﻿using System;
-using Microsoft.Extensions.Options;
+﻿using Microsoft.Extensions.Options;
 
-namespace ConsoleJson.Example
+namespace ConsoleJson.Example;
+
+public class ScopedService
 {
-    public class ScopedService
+    private readonly TransientFaultHandlingOptions _options;
+
+    public ScopedService(IOptionsSnapshot<TransientFaultHandlingOptions> options) =>
+        _options = options.Value;
+
+    public void DisplayValues()
     {
-        private readonly TransientFaultHandlingOptions _options;
-
-        public ScopedService(IOptionsSnapshot<TransientFaultHandlingOptions> options) =>
-            _options = options.Value;
-
-        public void DisplayValues()
-        {
-            Console.WriteLine($"TransientFaultHandlingOptions.Enabled={_options.Enabled}");
-            Console.WriteLine($"TransientFaultHandlingOptions.AutoRetryDelay={_options.AutoRetryDelay}");
-        }
+        Console.WriteLine($"TransientFaultHandlingOptions.Enabled={_options.Enabled}");
+        Console.WriteLine($"TransientFaultHandlingOptions.AutoRetryDelay={_options.AutoRetryDelay}");
     }
 }
