@@ -1,4 +1,3 @@
-﻿namespace CustomProvider.Example.Models
-{
-    public record Settings(string Id, string Value);
-}
+﻿namespace CustomProvider.Example.Models;
+
+public record Settings(string Id, string Value);
