@@ -1,15 +1,14 @@
 ﻿using Microsoft.Extensions.Configuration;
 
-namespace CustomProvider.Example.Providers
+namespace CustomProvider.Example.Providers;
+
+public class EntityConfigurationSource : IConfigurationSource
 {
-    public class EntityConfigurationSource : IConfigurationSource
-    {
-        private readonly string _connectionString;
+    private readonly string _connectionString;
 
-        public EntityConfigurationSource(string connectionString) =>
-            _connectionString = connectionString;
+    public EntityConfigurationSource(string connectionString) =>
+        _connectionString = connectionString;
 
-        public IConfigurationProvider Build(IConfigurationBuilder builder) =>
-            new EntityConfigurationProvider(_connectionString);
-    }
+    public IConfigurationProvider Build(IConfigurationBuilder builder) =>
+        new EntityConfigurationProvider(_connectionString);
 }
