@@ -1,22 +1,21 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace ExampleLibrary.Extensions.DependencyInjection
+namespace ExampleLibrary.Extensions.DependencyInjection;
+
+public static class ServiceCollectionExtensions
 {
-    public static class ServiceCollectionExtensions
+    public static IServiceCollection AddMyLibraryService(
+      this IServiceCollection services,
+      IConfiguration namedConfigurationSection)
     {
-        public static IServiceCollection AddMyLibraryService(
-          this IServiceCollection services,
-          IConfiguration namedConfigurationSection)
-        {
-            // Default library options are overridden
-            // by bound configuration values.
-            services.Configure<LibraryOptions>(namedConfigurationSection);
+        // Default library options are overridden
+        // by bound configuration values.
+        services.Configure<LibraryOptions>(namedConfigurationSection);
 
-            // Register lib services here...
-            // services.AddScoped<ILibraryService, DefaultLibraryService>();
+        // Register lib services here...
+        // services.AddScoped<ILibraryService, DefaultLibraryService>();
 
-            return services;
-        }
+        return services;
     }
 }
