@@ -1,4 +1,3 @@
-﻿namespace WorkerScope.Example
-{
-    record ObjectGraph(int Id, string Name, bool IsProcessed = false);
-}
+﻿namespace WorkerScope.Example;
+
+record ObjectGraph(int Id, string Name, bool IsProcessed = false);
