@@ -1,41 +1,35 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using WorkerServiceOptions.Example.Extensions;
+﻿using WorkerServiceOptions.Example.Extensions;
 
-namespace WorkerServiceOptions.Example
+namespace WorkerServiceOptions.Example;
+
+public class Worker : BackgroundService
 {
-    public class Worker : BackgroundService
+    private readonly ILogger<Worker> _logger;
+    private readonly PriorityQueue _priorityQueue;
+
+    public Worker(ILogger<Worker> logger, PriorityQueue priorityQueue) =>
+        (_logger, _priorityQueue) = (logger, priorityQueue);
+
+    protected override async Task ExecuteAsync(
+        CancellationToken stoppingToken)
     {
-        private readonly ILogger<Worker> _logger;
-        private readonly PriorityQueue _priorityQueue;
-
-        public Worker(ILogger<Worker> logger, PriorityQueue priorityQueue) =>
-            (_logger, _priorityQueue) = (logger, priorityQueue);
-
-        protected override async Task ExecuteAsync(
-            CancellationToken stoppingToken)
+        using IDisposable? scope = _logger.ProcessingWorkScope(DateTime.Now);
+        while (!stoppingToken.IsCancellationRequested)
         {
-            using IDisposable? scope = _logger.ProcessingWorkScope(DateTime.Now);
-            while (!stoppingToken.IsCancellationRequested)
+            WorkItem? nextItem = _priorityQueue.ProcessNextHighestPriority();
+            try
             {
-                WorkItem? nextItem = _priorityQueue.ProcessNextHighestPriority();
-                try
+                if (nextItem is not null)
                 {
-                    if (nextItem is not null)
-                    {
-                        _logger.PriorityItemProcessed(nextItem);
-                    }
+                    _logger.PriorityItemProcessed(nextItem);
                 }
-                catch (Exception ex)
-                {
-                    _logger.FailedToProcessWorkItem(ex);
-                }
-
-                await Task.Delay(1000, stoppingToken);
             }
+            catch (Exception ex)
+            {
+                _logger.FailedToProcessWorkItem(ex);
+            }
+
+            await Task.Delay(1000, stoppingToken);
         }
     }
 }
