@@ -1,11 +1,8 @@
-﻿using System;
+﻿namespace ConsoleDI.IEnumerableExample;
 
-namespace ConsoleDI.IEnumerableExample
+public class ConsoleMessageWriter : IMessageWriter
 {
-    public class ConsoleMessageWriter : IMessageWriter
-    {
-        public void Write(string message) => 
-            Console.WriteLine(
-                $"ConsoleMessageWriter.Write(message: \"{message}\")");
-    }
+    public void Write(string message) =>
+        Console.WriteLine(
+            $"ConsoleMessageWriter.Write(message: \"{message}\")");
 }
