@@ -1,24 +1,18 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.Hosting;
+﻿namespace DependencyInjection.Example;
 
-namespace DependencyInjection.Example
+public class Worker : BackgroundService
 {
-    public class Worker : BackgroundService
+    private readonly IMessageWriter _messageWriter;
+
+    public Worker(IMessageWriter messageWriter) =>
+        _messageWriter = messageWriter;
+
+    protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
-        private readonly IMessageWriter _messageWriter;
-
-        public Worker(IMessageWriter messageWriter) =>
-            _messageWriter = messageWriter;
-
-        protected override async Task ExecuteAsync(CancellationToken stoppingToken)
+        while (!stoppingToken.IsCancellationRequested)
         {
-            while (!stoppingToken.IsCancellationRequested)
-            {
-                _messageWriter.Write($"Worker running at: {DateTimeOffset.Now}");
-                await Task.Delay(1000, stoppingToken);
-            }
+            _messageWriter.Write($"Worker running at: {DateTimeOffset.Now}");
+            await Task.Delay(1000, stoppingToken);
         }
     }
 }
