@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace DependencyInjection.AntiPatterns
+﻿namespace DependencyInjection.AntiPatterns
 {
     public class ExampleDisposable : IDisposable
     {
