@@ -1,53 +1,45 @@
-﻿using System;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
+﻿namespace WorkerScope.Example;
 
-namespace WorkerScope.Example
+public class Worker : BackgroundService
 {
-    public class Worker : BackgroundService
+    private readonly ILogger<Worker> _logger;
+    private readonly IServiceScopeFactory _serviceScopeFactory;
+
+    public Worker(ILogger<Worker> logger, IServiceScopeFactory serviceScopeFactory) =>
+        (_logger, _serviceScopeFactory) = (logger, serviceScopeFactory);
+
+    protected override async Task ExecuteAsync(CancellationToken stoppingToken)
     {
-        private readonly ILogger<Worker> _logger;
-        private readonly IServiceScopeFactory _serviceScopeFactory;
-
-        public Worker(ILogger<Worker> logger, IServiceScopeFactory serviceScopeFactory) =>
-            (_logger, _serviceScopeFactory) = (logger, serviceScopeFactory);
-
-        protected override async Task ExecuteAsync(CancellationToken stoppingToken)
+        while (!stoppingToken.IsCancellationRequested)
         {
-            while (!stoppingToken.IsCancellationRequested)
+            using (IServiceScope scope = _serviceScopeFactory.CreateScope())
             {
-                using (IServiceScope scope = _serviceScopeFactory.CreateScope())
+                try
                 {
-                    try
-                    {
-                        _logger.LogInformation(
-                            "Starting scoped work, provider hash: {hash}.",
-                            scope.ServiceProvider.GetHashCode());
+                    _logger.LogInformation(
+                        "Starting scoped work, provider hash: {hash}.",
+                        scope.ServiceProvider.GetHashCode());
 
-                        var store = scope.ServiceProvider.GetRequiredService<IObjectStore>();
-                        var next = await store.GetNextAsync();
-                        _logger.LogInformation("{next}", next);
+                    var store = scope.ServiceProvider.GetRequiredService<IObjectStore>();
+                    var next = await store.GetNextAsync();
+                    _logger.LogInformation("{next}", next);
 
-                        var processor = scope.ServiceProvider.GetRequiredService<IObjectProcessor>();
-                        await processor.ProcessAsync(next);
-                        _logger.LogInformation("Processing {name}.", next.Name);
+                    var processor = scope.ServiceProvider.GetRequiredService<IObjectProcessor>();
+                    await processor.ProcessAsync(next);
+                    _logger.LogInformation("Processing {name}.", next.Name);
 
-                        var relay = scope.ServiceProvider.GetRequiredService<IObjectRelay>();
-                        await relay.RelayAsync(next);
-                        _logger.LogInformation("Processed results have been relayed.");
+                    var relay = scope.ServiceProvider.GetRequiredService<IObjectRelay>();
+                    await relay.RelayAsync(next);
+                    _logger.LogInformation("Processed results have been relayed.");
 
-                        var marked = await store.MarkAsync(next);
-                        _logger.LogInformation("Marked as processed: {next}", marked);
-                    }
-                    finally
-                    {
-                        _logger.LogInformation(
-                            "Finished scoped work, provider hash: {hash}.{nl}",
-                            scope.ServiceProvider.GetHashCode(), Environment.NewLine);
-                    }
+                    var marked = await store.MarkAsync(next);
+                    _logger.LogInformation("Marked as processed: {next}", marked);
+                }
+                finally
+                {
+                    _logger.LogInformation(
+                        "Finished scoped work, provider hash: {hash}.{nl}",
+                        scope.ServiceProvider.GetHashCode(), Environment.NewLine);
                 }
             }
         }
