@@ -1,18 +1,13 @@
-﻿using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
+﻿namespace WorkerServiceOptions.Example;
 
-namespace WorkerServiceOptions.Example
+class Program
 {
-    class Program
-    {
-        static Task Main(string[] args) =>
-            CreateHostBuilder(args).Build().RunAsync();
+    static Task Main(string[] args) =>
+        CreateHostBuilder(args).Build().RunAsync();
 
-        static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureServices((_, services) =>
-                    services.AddHostedService<Worker>()
-                            .AddTransient<PriorityQueue>());
-    }
+    static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .ConfigureServices((_, services) =>
+                services.AddHostedService<Worker>()
+                        .AddTransient<PriorityQueue>());
 }
