@@ -1,17 +1,12 @@
-﻿using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
+﻿namespace WorkerService.Example;
 
-namespace WorkerService.Example
+class Program
 {
-    class Program
-    {
-        static Task Main(string[] args) =>
-            CreateHostBuilder(args).Build().RunAsync();
+    static Task Main(string[] args) =>
+        CreateHostBuilder(args).Build().RunAsync();
 
-        static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureServices((_, services) =>
-                    services.AddHostedService<Worker>());
-    }
+    static IHostBuilder CreateHostBuilder(string[] args) =>
+        Host.CreateDefaultBuilder(args)
+            .ConfigureServices((_, services) =>
+                services.AddHostedService<Worker>());
 }
